+++ www/src/components/page-with-plugin-searchbar.js
@@ -4,7 +4,7 @@ import { Fragment } from "react"
 import loadable from "@loadable/component"
 import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 
-const PluginSearchBar = loadable(() => import("./plugin-searchbar-body"))
+const PluginSearchBar = loadable(() => import(`./plugin-searchbar-body`))
 
 const PageWithPluginSearchBar = ({ isPluginsIndex, location, children }) => (
   <Fragment>