+++ packages/gatsby-plugin-benchmark-reporting/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-benchmark-reporting@0.2.4...gatsby-plugin-benchmark-reporting@0.2.5) (2020-05-20)
+
+**Note:** Version bump only for package gatsby-plugin-benchmark-reporting
+
 ## [0.2.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-benchmark-reporting@0.2.3...gatsby-plugin-benchmark-reporting@0.2.4) (2020-05-13)
 
 ### Features